# Conflicts:
#	identityprovider/src/main/java/nz/ac/canterbury/seng302/identityprovider/database/UserRepository.java
#	identityprovider/src/main/java/nz/ac/canterbury/seng302/identityprovider/service/RegisterServerService.java
#	portfolio/subdirectory/demodb.mv.db
#	portfolio/subdirectory/demodb.trace.db

+++ identityprovider/src/main/java/nz/ac/canterbury/seng302/identityprovider/database/UserRepository.java
@@ -1,12 +1,11 @@
 package nz.ac.canterbury.seng302.identityprovider.database;
 
-import java.util.List;
-
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserRepository extends CrudRepository<UserModel, Integer> {
+public interface UserRepository extends CrudRepository<UserModel, Long> {
 
-    UserModel findByUsername(String username);
+  UserModel findByUsername(String username);
 
-    UserModel findById(int id);
+  UserModel findById(long id);
 }
